refactor: reorganize project structure and standardize outputs
- Move source files to src/late_chunking/
- Create modular directory structure
- Move data files to data/
- Standardize output locations to outputs/<example>/
- Update paths in all files
- Add comprehensive documentation
