--- templates/visualizer.html
@@ -1,192 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Embedding Space Visualization</title>
-    <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            margin: 0;
-            padding: 20px;
-            background-color: #f5f5f5;
-        }
-        .container {
-            max-width: 1400px;
-            margin: 0 auto;
-            background-color: white;
-            padding: 20px;
-            border-radius: 8px;
-            box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-        }
-        h1 {
-            color: #333;
-            text-align: center;
-            margin-bottom: 30px;
-        }
-        .plot-container {
-            display: flex;
-            flex-wrap: wrap;
-            gap: 20px;
-            justify-content: center;
-        }
-        .plot {
-            flex: 1;
-            min-width: 600px;
-            background-color: white;
-            padding: 15px;
-            border-radius: 8px;
-            box-shadow: 0 1px 3px rgba(0,0,0,0.1);
-        }
-        .hover-info {
-            margin-top: 10px;
-            padding: 10px;
-            min-height: 100px;
-            background-color: #f8f9fa;
-            border-radius: 4px;
-            border: 1px solid #dee2e6;
-            font-size: 14px;
-            line-height: 1.5;
-            white-space: pre-wrap;
-            overflow-y: auto;
-            max-height: 200px;
-        }
-        .loading {
-            text-align: center;
-            padding: 50px;
-            font-size: 18px;
-            color: #666;
-        }
-        .error {
-            color: #d32f2f;
-            text-align: center;
-            padding: 20px;
-            background-color: #ffebee;
-            border-radius: 4px;
-            margin: 20px 0;
-            display: none;
-        }
-        .plot-title {
-            font-size: 16px;
-            font-weight: bold;
-            margin-bottom: 10px;
-            color: #333;
-        }
-    </style>
-</head>
-<body>
-    <div class="header">
-        <h1>Embedding Space Visualization</h1>
-        <div class="controls">
-            <label for="model-select">Embedding Model:</label>
-            <select id="model-select" onchange="loadVisualizations()">
-                {% for model in models %}
-                <option value="{{ model }}">{{ model }}</option>
-                {% endfor %}
-            </select>
-        </div>
-    </div>
-    
-    <div class="container">
-        <div id="error" class="error"></div>
-        <div id="loading" class="loading">Loading visualizations...</div>
-        <div class="plot-container">
-            <div class="plot">
-                <div class="plot-title">Late Chunking Embeddings</div>
-                <div id="late-chunking-plot"></div>
-                <div id="late-chunking-hover" class="hover-info">Hover over points to see the text content...</div>
-            </div>
-            <div class="plot">
-                <div class="plot-title">Traditional Chunking Embeddings</div>
-                <div id="traditional-plot"></div>
-                <div id="traditional-hover" class="hover-info">Hover over points to see the text content...</div>
-            </div>
-        </div>
-    </div>
-
-    <script>
-        function updateHoverInfo(hoverDiv, data) {
-            if (data) {
-                const text = data.points[0].customdata[0];
-                const docId = data.points[0].data.name || 'Unknown Document';
-                hoverDiv.innerHTML = `<strong>Document:</strong> ${docId}\n<strong>Text:</strong>\n${text}`;
-            }
-        }
-
-        function loadVisualizations() {
-            const modelSelect = document.getElementById('model-select');
-            const selectedModel = modelSelect.value;
-            
-            fetch(`/get_visualizations?model=${selectedModel}`)
-                .then(response => response.json())
-                .then(data => {
-                    if (data.error) {
-                        showError(data.error);
-                    } else {
-                        // Hide loading message
-                        document.getElementById('loading').style.display = 'none';
-                        
-                        // Configure hover modes
-                        const config = {
-                            displayModeBar: true,
-                            scrollZoom: true
-                        };
-
-                        // Update layout to disable hover labels
-                        const layoutUpdate = {
-                            hovermode: 'closest',
-                            hoverlabel: { enabled: false }
-                        };
-
-                        // Merge layout updates
-                        data.late_chunking.layout = {
-                            ...data.late_chunking.layout,
-                            ...layoutUpdate
-                        };
-                        data.traditional.layout = {
-                            ...data.traditional.layout,
-                            ...layoutUpdate
-                        };
-                        
-                        // Create plots
-                        Plotly.newPlot('late-chunking-plot', 
-                            data.late_chunking.data,
-                            data.late_chunking.layout,
-                            config
-                        );
-                        
-                        Plotly.newPlot('traditional-plot',
-                            data.traditional.data,
-                            data.traditional.layout,
-                            config
-                        );
-
-                        // Add hover event listeners
-                        const lateChunkingHover = document.getElementById('late-chunking-hover');
-                        const traditionalHover = document.getElementById('traditional-hover');
-
-                        document.getElementById('late-chunking-plot').on('plotly_hover', function(data) {
-                            updateHoverInfo(lateChunkingHover, data);
-                        });
-
-                        document.getElementById('traditional-plot').on('plotly_hover', function(data) {
-                            updateHoverInfo(traditionalHover, data);
-                        });
-                    }
-                })
-                .catch(error => {
-                    showError('Error loading visualizations: ' + error.message);
-                });
-        }
-
-        function showError(message) {
-            const errorDiv = document.getElementById('error');
-            errorDiv.textContent = message;
-            errorDiv.style.display = 'block';
-            document.getElementById('loading').style.display = 'none';
-        }
-
-        // Initial load
-        loadVisualizations();
-    </script>
-</body>
-</html>
